Conflicts:
	src/main/java/evg/testt/controller/DepartmentController.java
	src/main/resources/spring/dispatcherServlet.xml

+++ src/main/java/evg/testt/controller/DepartmentController.java
@@ -1,6 +1,8 @@
 package evg.testt.controller;
 
 import evg.testt.model.Department;
+import evg.testt.model.User;
+import evg.testt.service.ChatService;
 import evg.testt.service.DepartmentService;
 import evg.testt.util.JspPath;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +11,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.sql.SQLException;
 
 @Controller
@@ -17,55 +22,53 @@ public class DepartmentController{
     @Autowired
     DepartmentService departmentService;
 
+    @Autowired
+    ChatService chatService;
+
     @RequestMapping(value = "/dep", method = RequestMethod.GET)
-    public ModelAndView showAll() throws SQLException {
-        return new ModelAndView(JspPath.DEPARTMENT_ALL, "departments", departmentService.getAll());
+    public ModelAndView showAll(HttpServletRequest request) throws SQLException {
+        HttpSession session = request.getSession();
+        User sessionUser = (User) session.getAttribute("user");
+        ModelAndView modelAndView = new ModelAndView(JspPath.DEPARTMENT_ALL);
+        modelAndView.addObject("email", sessionUser.getEmail());
+        modelAndView.addObject("departments", departmentService.getAll());
+        return modelAndView;
     }
 
     @RequestMapping(value = "/depSaveOrUpdate", method = RequestMethod.POST)
-    public String addNewOne(@RequestParam(required = false) Integer id, @RequestParam(required = true) String name) {
+    public String addNewOne(@RequestParam(required = false) Integer id, @RequestParam(required = true) String name) throws SQLException {
         Department department = Department.newBuilder().setName(name).setId(id).build();
         if(id==null){
-            try {
-                departmentService.insert(department);
-            }
-            catch (SQLException e) {
-            }
+            departmentService.insert(department);
+        }else{
+            departmentService.update(department);
         }
-        else
-            {
-            try {
-                departmentService.update(department);
-            } catch (SQLException e) {
-
-            }
-        }
-
-
         return "redirect:/dep";
     }
 
     @RequestMapping(value = "/depDelete", method = RequestMethod.POST)
-    public String deleteOne(@RequestParam(required = true) Integer id) {
-        Department department = Department.newBuilder().setId(id).build();
-        try {
+    public String deleteOne(@RequestParam(required = true) Integer id) throws SQLException {
+        String direct = "";
+        Department department; //= Department.newBuilder().setId(id).build();
+        department = departmentService.getById(id);
+        if (department.getEmployees().size() == 0) {
             departmentService.delete(department);
-        } catch (SQLException e) {
+            direct = "redirect:/dep";
+        } else {
+            direct = "redirect:/employees?id="+department.getId()+"&toDel=yes";
         }
-        return "redirect:/dep";
+
+        return direct;
     }
 
     @RequestMapping(value = "/depEdit", method = RequestMethod.POST)
-    public ModelAndView updateOne(@RequestParam(required = false) Integer id) {
+    public ModelAndView updateOne(@RequestParam(required = false) Integer id) throws SQLException {
         ModelAndView modelAndView = new ModelAndView(JspPath.DEPARTMENT_EDIT);
         if(id!=null){
-            Department department = null;
-            try {
-                department = departmentService.getById(id);
-                modelAndView.addObject("department", department);
-            } catch (SQLException e) {
-            }
+            Department department = departmentService.getById(id);
+            modelAndView.addObject("department", department);
         }
         return modelAndView;
     }
+
 }
